Core: Additional Validity Check in Breakpoint Check
-Problem was detected when setting and removing breakpoints. It was
possible to remove a breakpoint while the main loop was checking to see
if the current memoryPointer matched any of the breakpoint values.

This caused the loop to index nonexistent values causing an exception.
Now each if has a check in addition so that it can stop if there is no
such index.