AssemblyView: Run Button also enables Step
Previously if the window ran to a breakpoint and pressed run again it
would not move. It would stay stuck at the same breakpoint. This is
because this memorypointer value still registers as a breakpoint and
sees no reason to  skip. To combat this run also enables Step so that it
will skip one break point then run.