Core: Fixed bug causing first breakpoint to not Work
-To fix the issue of indexing a non-existent breakpoint a boolean that
checks if the length is large enough for the index to exist. Either the
count had to be reduced by one or the index had to be padded by 1, but
both were done. This was a mistake that caused the first index to be
ignored entirely.

This has been fixed by removing both and simplifying the logic as a
whole.