GameWindow: Fixed Bug Causing Purple/Invisible Colors
In the GPU class is the Gameboy Palette array of colors that make up
what will be the DMG color shades. The array of colors that make up the
screens are populated with one of those colors.

The code that set the screen to those colors used a color from the
palette. When a color would get modified in the screen the colors in the
palette would change as well because the colors are not being copied,
they are becoming references of the palette colors.

Now the new copy constructor shall be used to copy the values of another
color object while not changing the existing one.