PaletteEditor: Added Function to Save Current Palettes to Palettes.XML
The Save all To File button will now take all the current palettes and
save them to the XML as they are in memory. It will make a copy of the
file before making changes if there is no existing copy, and it will
write over the current file if there is one.