Moved Halt Bug Variable and Implementation to Core
Previously the implementation of the halt bug was contained within the
CPU class. While this worked functionally, it was not ideal for
readability. The implementation was tucked away in the select opcode
function which is non-obvious. Instead, I have made it much more clear
by moving it to the core driver function after the select Opcode
function call. This is a more obvious and sensical location and is
easier to remember.