Simplified and Optimized checkForInterrupt()
Previously a loop was utilized to check each bit 1 by 1 for an
interrupt. This is flawed for two reasons. Immediately because not all 8
bits are responsible for a pending interrupt. 0x1-0x1F are interrupts,
but the rest are not. Therefore this could lead to a false interrupt
detection.

Secondly, because it is quite a lot more work for a relatively simple
task. The downside is that it was potentially easier to read before, so
I have added comments to avoid confusion.