Improved/Fixed Halt Instruction 76 and Interrupt Logic
Previously halt was not implemented correctly. It ignored the different
potential halt states. One where halt mode is entered correctly, the
next where halt mode is entered, but the interrupt jump is not made, and
one where halt mode is not entered and instead a halt bug occurs.

The logic for the halt instruction now passes the Blargg instruction
test 2.