Updating Screen is now done immediately
Early on in the project the performance was quite inferior. One way that
was helped was by moving the quite resource expensive screen updating to
it's own task so that the emulator would continue running while the
screen drew.

This was a big help for performance, but let to a worse experience. If
the screen is updated immediately it is guaranteed to draw the exact
frame as intended. If it is a separate task then it is possible that the
screen data will have changed by the time it is done drawing causing
visual artifacts.

Now that performance is no longer a major issue I have removed the Task
code and made it call the function directly.