Updated drawFullScreenMaps() to return SFML::Image
Previously this function was designed to write the full window to an
Image then draw it to the screen to display it. This is useful, but very
specific and can't be re-used.

Instead, we now simply return the Image so that other code can choose
what is done with it instead of automatically draw it.