GameWindow: Fix for ScrollX dependent scenes (Huds)
The gameboy is meant to draw one line at a time. Each line can have
their own individual scrollX value attached for effects. Previously I
ignored this and just drew the whole screen with the scrollX that was
set when I was ready to draw the frame. This was very inaccurate.

Now I record the scrollX value when each line is drawn so that I can use
it when drawing the full frame to get the scroll that was intended.

This is not perfect and more changes will need to be made. This is
because some games can require areas to get redrawn after they are put
onto screen. Therefore I need to draw lines to the next frame when it is
meant to, then draw the full frame to the next screen, instead of
drawing the full screen after those changes are made.