Memory: Added returnSaveDataFromMemory()
This function returns the current data in RAM. This data will be in the
form of an array. With this return data it can be used to save as a file
that can be loaded back in for saving functionality.

This function just gets the raw data so that another function can decide
how it will be saved.