Memory: Added createSaveFile(path, overwrite)
Previously a function was created to return the save data from RAM as an
array. While this is a useful function, there was no functionality
making use of it. This function can take that array and save it as a
.sav file.

If there is no existing .sav file or one has chosen to overwrite it then
the save data bytes will be saved as an .sav in the with the same name
and in the same location as the rom file.