MainWindow: When Rom Fails to Load romFilePath is set to emptry string
If one loads a game, then attempts to load another game and it fails,
the first game's rompath will still be in memory. This can lead to odd
results like saving a .sav file for a game you no longer have loaded.
Now the romFilePath is set to "" when rom fails to load.