MainWindow: Improvements to Thread Termination
The emulator thread is created when a file is loaded to run the game.
This thread sometimes gives errors when closing the emulator.

To prevent this multiple changes have been made to give the thread a
chance to terminate safely. If it does not within a certain time it will
terminate regardless.