Memory: createSaveFile now checks cartridgeType
Previously ramEnabled was mistakenly used to determine if the rom used
ram and therefore a .sav file could be generated.

This is a sign that the variable name is too general and should likely
be documented better within the program. Instead, the cartridge type is
now used to determine if the rom will make use of ram.