Rom: Array containing rom data is now a fixed length
Previously romData was set to the array that was loaded in, regardless
of it's length. This is a problem for when the programming of a game
expects to read past this length. This may happen during rom banking.
The rest of the data would be zeros, but the code may still need to
write a full bank.

The rom data is now the same fixed length it used to be.