CPU: Fixed executeStop()
This was meant to check if a joypad interrupt was enabled and end Stop
mode if it is. But code that was specifically for handling interrupts
was left in accidentally. That has now been removed so that it does not
effect emulation.

It is now purely to end Stop mode as intended.