Configured DLLMaps to find csfml DLLs on Linux
Previously on linux with Mono the SFMLNet DLLs would search for "csfml-graphics-2" and equivalently named DLLs for each of the other SFML DLL types.

Since the installed files are not named in this format a DLLMap has been added to allow it to search for the actual 2.4 CSFML linux library files.