Can now run several roms consecutively properly
Previously if one loaded a rom, then selected another rom to load in, it was likely that the new rom would not emulate properly.

This was due to not resetting the emulation state before starting the second rom. If the processor was in the middle of an interrupt, or the gpu was in a particular rendering mode that the rom is not expecting to be in when it first starts it could lead to very unexpected results.

To resolve this, each emulation class has a reset function where applicable so that the emulator will have a clean slate when running the next rom.